pyproject.toml
[project]
name = 'pension-planning-agent'
version = "0.0.0"
description = "AI Agent for personalized pension planning, offering savings projections, retirement income analysis, and contribution optimization for long-term financial security."
authors = [{name='Petr Vanek', email= 'petrr.vanekk@gmail.com'}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "httpx>=0.28.1",
    "litellm>=1.0.0",
    "logfire>=3.4.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "streamlit>=1.41.1",
]

[project.urls]
repository = "https://github.com/VanekPetr/pension-planning-agent"

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "*__init__.py"
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pension_planning_agent"]

[tool.hatch.build]
include = [
    "LICENSE",    # Ensure the LICENSE file is included in your package
    "README.md",
    "src/pension_planning_agent"
]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
